Add messagebroker service using RabbitMQ
This commit introduces a new `messagebroker` service in both `docker-compose.override.yml` and `docker-compose.yml`. The service utilizes the `rabbitmq:management` image and is configured with default user credentials. It exposes ports 5672 and 15672 for communication and management, respectively, and is set to restart always with a specified container name and hostname.